Add files via upload
This is project done for assignment by using development technologies for creating authentication app. Which calculates the user attendance comes in to page and login and login. it also counts the no of time login by any user and logout. it gives the surity upon the presence of employee active on the job portal.